+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add location field to experiment upload form.  [#259](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/259)
+
 ### Changed
 
 - Remove add crop logic from experiment upload form.  [#235](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/235)